Added saferm but slow implementation of FixParser
Added perf test to compare FixReader and NormalReader.
Current slowdown is aprox 10 times, but ahrd to meassure because:
Currently the only working combination FixReader + big-old-broken is pushed
Other combinations fails on not well formed document, which is weird, as
those are our existing data

TestRunner aligned to Xmx 4g, to enable parsin g of that 800mb file
The xmx is declared in report-jtreg-lib, but it seems idea is taking it
from parent, and thsu seting 762mb *behind* the 4gb. To workaround it,
go in ideas's settings->Build...->Build Tools->RunningTests and disable
three checkobxes.

Maybe this can be done better by seeting the shurefire-settings to
parent pom or who knows where...